+++ cmd/mdatagen/internal/samplereceiver/metrics_test.go
@@ -9,8 +9,10 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
+	"go.opentelemetry.io/otel/sdk/metric/metricdata/metricdatatest"
 
 	"go.opentelemetry.io/collector/cmd/mdatagen/internal/samplereceiver/internal/metadata"
+	"go.opentelemetry.io/collector/cmd/mdatagen/internal/samplereceiver/internal/metadatatest"
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/consumer/consumertest"
 	"go.opentelemetry.io/collector/receiver/receivertest"
@@ -24,11 +26,11 @@ func TestGeneratedMetrics(t *testing.T) {
 }
 
 func TestComponentTelemetry(t *testing.T) {
-	tt := setupTestTelemetry()
+	tt := metadatatest.SetupTelemetry()
 	factory := NewFactory()
 	receiver, err := factory.CreateMetrics(context.Background(), tt.NewSettings(), componenttest.NewNopHost(), new(consumertest.MetricsSink))
 	require.NoError(t, err)
-	tt.assertMetrics(t, []metricdata.Metrics{
+	tt.AssertMetrics(t, []metricdata.Metrics{
 		{
 			Name:        "otelcol_batch_size_trigger_send",
 			Description: "Number of times the batch was sent due to a size trigger [deprecated since v0.110.0]",
@@ -57,11 +59,11 @@ func TestComponentTelemetry(t *testing.T) {
 				},
 			},
 		},
-	})
+	}, metricdatatest.IgnoreTimestamp())
 	rcv, ok := receiver.(nopReceiver)
 	require.True(t, ok)
 	rcv.initOptionalMetric()
-	tt.assertMetrics(t, []metricdata.Metrics{
+	tt.AssertMetrics(t, []metricdata.Metrics{
 		{
 			Name:        "otelcol_batch_size_trigger_send",
 			Description: "Number of times the batch was sent due to a size trigger [deprecated since v0.110.0]",
@@ -102,6 +104,6 @@ func TestComponentTelemetry(t *testing.T) {
 				},
 			},
 		},
-	})
+	}, metricdatatest.IgnoreTimestamp())
 	require.NoError(t, tt.Shutdown(context.Background()))
 }