+++ exporter/exporterhelper/internal/queue_sender_test.go
@@ -23,14 +23,14 @@ import (
 	"go.opentelemetry.io/collector/exporter/exporterqueue"
 	"go.opentelemetry.io/collector/exporter/exportertest"
 	"go.opentelemetry.io/collector/exporter/internal"
-	"go.opentelemetry.io/collector/exporter/internal/queue"
+	"go.opentelemetry.io/collector/exporter/internal/storagetest"
 	"go.opentelemetry.io/collector/pipeline"
 )
 
 func TestQueuedRetry_StopWhileWaiting(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
+			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			qCfg := NewDefaultQueueConfig()
 			qCfg.NumConsumers = 1
 			rCfg := configretry.NewDefaultBackOffConfig()
@@ -192,7 +192,8 @@ func TestQueuedRetryHappyPath(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool, tt struct {
 		name         string
 		queueOptions []Option
-	}) {
+	},
+	) {
 		t.Run(testName, func(t *testing.T) {
 			resetFeatureGate := setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
 			tel, err := componenttest.SetupTelemetry(defaultID)
@@ -244,7 +245,7 @@ func TestQueuedRetryHappyPath(t *testing.T) {
 func TestQueuedRetry_QueueMetricsReported(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
+			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			dataTypes := []pipeline.Signal{pipeline.SignalLogs, pipeline.SignalTraces, pipeline.SignalMetrics}
 			for _, dataType := range dataTypes {
 				tt, err := componenttest.SetupTelemetry(defaultID)
@@ -283,7 +284,7 @@ func TestQueuedRetry_QueueMetricsReported(t *testing.T) {
 func TestNoCancellationContext(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
+			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			deadline := time.Now().Add(1 * time.Second)
 			ctx, cancelFunc := context.WithDeadline(context.Background(), deadline)
 			cancelFunc()
@@ -307,7 +308,7 @@ func TestNoCancellationContext(t *testing.T) {
 func TestQueueConfig_Validate(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
+			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			qCfg := NewDefaultQueueConfig()
 			require.NoError(t, qCfg.Validate())
 
@@ -361,9 +362,10 @@ func TestQueueRetryWithDisabledQueue(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool, tt struct {
 		name         string
 		queueOptions []Option
-	}) {
+	},
+	) {
 		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
+			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			set := exportertest.NewNopSettings()
 			logger, observed := observer.New(zap.ErrorLevel)
 			set.Logger = zap.New(logger)
@@ -393,7 +395,7 @@ func TestQueueRetryWithDisabledQueue(t *testing.T) {
 func TestQueueFailedRequestDropped(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
+			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			set := exportertest.NewNopSettings()
 			logger, observed := observer.New(zap.ErrorLevel)
 			set.Logger = zap.New(logger)
@@ -417,7 +419,7 @@ func TestQueueFailedRequestDropped(t *testing.T) {
 func TestQueuedRetryPersistenceEnabled(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
+			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			tt, err := componenttest.SetupTelemetry(defaultID)
 			require.NoError(t, err)
 			t.Cleanup(func() { require.NoError(t, tt.Shutdown(context.Background())) })
@@ -432,8 +434,8 @@ func TestQueuedRetryPersistenceEnabled(t *testing.T) {
 				WithRetry(rCfg), WithQueue(qCfg))
 			require.NoError(t, err)
 
-			var extensions = map[component.ID]component.Component{
-				storageID: queue.NewMockStorageExtension(nil),
+			extensions := map[component.ID]component.Component{
+				storageID: storagetest.NewMockStorageExtension(nil),
 			}
 			host := &MockHost{Ext: extensions}
 
@@ -450,7 +452,7 @@ func TestQueuedRetryPersistenceEnabled(t *testing.T) {
 func TestQueuedRetryPersistenceEnabledStorageError(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
+			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			storageError := errors.New("could not get storage client")
 			tt, err := componenttest.SetupTelemetry(defaultID)
 			require.NoError(t, err)
@@ -465,8 +467,8 @@ func TestQueuedRetryPersistenceEnabledStorageError(t *testing.T) {
 				WithUnmarshaler(mockRequestUnmarshaler(&mockRequest{})), WithRetry(rCfg), WithQueue(qCfg))
 			require.NoError(t, err)
 
-			var extensions = map[component.ID]component.Component{
-				storageID: queue.NewMockStorageExtension(storageError),
+			extensions := map[component.ID]component.Component{
+				storageID: storagetest.NewMockStorageExtension(storageError),
 			}
 			host := &MockHost{Ext: extensions}
 
@@ -497,8 +499,8 @@ func TestQueuedRetryPersistentEnabled_NoDataLossOnShutdown(t *testing.T) {
 				WithUnmarshaler(mockRequestUnmarshaler(mockReq)), WithRetry(rCfg), WithQueue(qCfg))
 			require.NoError(t, err)
 
-			var extensions = map[component.ID]component.Component{
-				storageID: queue.NewMockStorageExtension(nil),
+			extensions := map[component.ID]component.Component{
+				storageID: storagetest.NewMockStorageExtension(nil),
 			}
 			host := &MockHost{Ext: extensions}
 
@@ -537,12 +539,18 @@ func TestQueuedRetryPersistentEnabled_NoDataLossOnShutdown(t *testing.T) {
 func TestQueueSenderNoStartShutdown(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
-			queue := queue.NewBoundedMemoryQueue[internal.Request](queue.MemoryQueueSettings[internal.Request]{})
+			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			set := exportertest.NewNopSettings()
+			queue := exporterqueue.NewMemoryQueueFactory[internal.Request]()(
+				context.Background(),
+				exporterqueue.Settings{
+					Signal:           pipeline.SignalTraces,
+					ExporterSettings: set,
+				},
+				exporterqueue.NewDefaultConfig())
 			obsrep, err := NewExporter(ObsReportSettings{
 				ExporterID:             exporterID,
-				ExporterCreateSettings: exportertest.NewNopSettings(),
+				ExporterCreateSettings: set,
 			})
 			require.NoError(t, err)
 			qs := NewQueueSender(queue, set, 1, "", obsrep, exporterbatcher.NewDefaultConfig())