+++ scraper/scraperhelper/doc.go
@@ -1,0 +1,7 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+//go:generate mdatagen metadata.yaml
+
+// Package scraperhelper provides utilities for scrapers.
+package scraperhelper // import "go.opentelemetry.io/collector/scraper/scraperhelper"